+++ boot-rabbitmq-thymeleaf/src/test/java/com/poc/boot/rabbitmq/common/ContainerConfiguration.java
@@ -1,0 +1,19 @@
+package com.poc.boot.rabbitmq.common;
+
+import org.springframework.boot.devtools.restart.RestartScope;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.testcontainers.containers.RabbitMQContainer;
+import org.testcontainers.utility.DockerImageName;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class ContainerConfiguration {
+
+    @Bean
+    @ServiceConnection
+    @RestartScope
+    RabbitMQContainer rabbitMQContainer() {
+        return new RabbitMQContainer(DockerImageName.parse("rabbitmq").withTag("4.0.5-management"));
+    }
+}
