+++ boot-rabbitmq-thymeleaf/src/test/java/com/poc/boot/rabbitmq/controller/MessageControllerTest.java
@@ -19,8 +19,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(MessageController.class)
@@ -29,7 +29,7 @@ class MessageControllerTest {
 
     @Autowired private MockMvc mockMvc;
 
-    @MockBean private OrderMessageSender orderMessageSender;
+    @MockitoBean private OrderMessageSender orderMessageSender;
 
     @Test
     void handleMessage() throws Exception {
