+++ jpa/boot-data-envers/src/test/java/com/example/envers/common/ContainersConfig.java
@@ -12,6 +12,6 @@ public class ContainersConfig {
     @Bean
     @ServiceConnection
     PostgreSQLContainer<?> postgreSQLContainer() {
-        return new PostgreSQLContainer<>(DockerImageName.parse("postgres:17.2-alpine"));
+        return new PostgreSQLContainer<>(DockerImageName.parse("postgres").withTag("17.2-alpine"));
     }
 }
