+++ jpa/boot-data-envers/src/test/java/com/example/envers/repository/SchemaValidationTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-@DataJpaTest(properties = {"spring.jpa.hibernate.ddl-auto=validate", "spring.test.database.replace=none"})
+@DataJpaTest(properties = {"spring.jpa.hibernate.ddl-auto=validate"})
 @Import(ContainersConfig.class)
 class SchemaValidationTest {
