+++ scheduler/boot-scheduler-quartz/src/main/resources/application.properties
@@ -39,12 +39,12 @@ spring.quartz.properties.org.quartz.scheduler.instanceId=AUTO
 spring.quartz.properties.org.quartz.jobStore.driverDelegateClass=org.quartz.impl.jdbcjobstore.PostgreSQLDelegate
 ## Default is RamJobStore, if you want to use JDBC Job Store, you need to set the following properties
 spring.quartz.properties.org.quartz.jobStore.class=org.quartz.impl.jdbcjobstore.JobStoreTX
-spring.quartz.properties.org.quartz.jobStore.dataSource=quartzDataSource
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.provider=hikaricp
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.driver=${spring.datasource.driver-class-name}
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.URL=${spring.datasource.url}
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.user=${spring.datasource.username}
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.password=${spring.datasource.password}
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.maxConnections=10
-spring.quartz.properties.org.quartz.dataSource.quartzDataSource.validationQuery=select 1
+spring.quartz.properties.org.quartz.jobStore.dataSource=quartzDS
+spring.quartz.properties.org.quartz.dataSource.quartzDS.provider=hikaricp
+spring.quartz.properties.org.quartz.dataSource.quartzDS.driver=${spring.datasource.driver-class-name}
+spring.quartz.properties.org.quartz.dataSource.quartzDS.URL=${spring.datasource.url}
+spring.quartz.properties.org.quartz.dataSource.quartzDS.user=${spring.datasource.username}
+spring.quartz.properties.org.quartz.dataSource.quartzDS.password=${spring.datasource.password}
+spring.quartz.properties.org.quartz.dataSource.quartzDS.maxConnections=10
+spring.quartz.properties.org.quartz.dataSource.quartzDS.validationQuery=select 1
